requirements.txt
autopep8
flake8
mysql-connector-python==8.0.26
pandas==1.3.3
imap-tools==0.48.0
sqlalchemy==1.4.23
pymysql==1.0.2
beautifulsoup4==4.10.0
lxml==4.6.3
requests==2.26.0
fasttext==0.9.2
spacy==3.1.3
google-auth==2.0.0
google-auth-oauthlib==0.4.6
google-cloud-language
deepl
googlemaps
nltk
pattern
